requirements.txt
altgraph==0.17.4
et_xmlfile==2.0.0
lxml==5.3.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
pywin32-ctypes==0.2.3
setuptools==75.8.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
